split reactpanel, extension public API mvp
