+++ package.json
@@ -2,7 +2,7 @@
   "name": "ethcode",
   "displayName": "ETHcode",
   "description": "Ethereum IDE for VS Code",
-  "version": "0.1.15",
+  "version": "0.2.2",
   "publisher": "Ethential",
   "categories": [
     "Debuggers",
